feat: added authorization to the comments management

+++ server/Controllers/comment.cs
@@ -2,17 +2,21 @@ using Microsoft.AspNetCore.Mvc;
 using server.Services;
 using server.Models;
 using MongoDB.Driver;
+using System.Security.Claims;
+using Microsoft.Extensions.Options;
+using server.DTOs;
 
 [ApiController]
 [Route("api/[controller]")]
 public class CommentController : ControllerBase
 {
-
   private readonly MongoDBService _mongoDBService;
+  private readonly JwtSettings _jwtSettings;
 
-  public CommentController(MongoDBService mongoDBService)
+  public CommentController(MongoDBService mongoDBService, IOptions<JwtSettings> jwtSettings)
   {
     _mongoDBService = mongoDBService;
+    _jwtSettings = jwtSettings.Value;
   }
 
   [HttpGet]
@@ -29,12 +33,21 @@ public class CommentController : ControllerBase
     }
   }
 
-  [HttpGet("customer/{customerId:length(24)}")]
-  public async Task<IActionResult> GetCustomerSpecificComments(string customerId)
+  [HttpGet("customer")]
+  public async Task<IActionResult> GetCustomerSpecificComments()
   {
     try
     {
+      var customerId = await AuthorizeCustomerAsync();
+      if (customerId == null)
+      {
+        return Unauthorized("Unauthorised User");
+      }
       var comments = await _mongoDBService.GetCustomerSpecificCommentsAsync(customerId);
+      if (comments.Count == 0)
+      {
+        return NotFound(new { Message = "There aren't any comments to show!!!" });
+      }
       return Ok(comments);
     }
     catch (Exception error)
@@ -43,12 +56,21 @@ public class CommentController : ControllerBase
     }
   }
 
-  [HttpGet("vender/{venderId:length(24)}")]
-  public async Task<IActionResult> GetVenderSpecificComments(string venderId)
+  [HttpGet("vender")]
+  public async Task<IActionResult> GetVenderSpecificComments()
   {
     try
     {
+      var venderId = await AuthorizeVenderAsync();
+      if (venderId == null)
+      {
+        return Unauthorized("Unauthorised User");
+      }
       var comments = await _mongoDBService.GetVenderSpecificCommentsAsync(venderId);
+      if (comments.Count == 0)
+      {
+        return NotFound(new { Message = "There aren't any comments to show!!!" });
+      }
       return Ok(comments);
     }
     catch (Exception error)
@@ -82,11 +104,25 @@ public class CommentController : ControllerBase
     }
   }
 
-  [HttpPost]
-  public async Task<IActionResult> Post([FromBody] Comments comment)
+  [HttpPost("{venderId:length(24)}")]
+  public async Task<IActionResult> Post(string venderId, [FromBody] CommentDto commentDto)
   {
     try
     {
+      var customerId = await AuthorizeCustomerAsync();
+      if (customerId == null)
+      {
+        return Unauthorized("Unauthorised User");
+      }
+
+      var comment = new Comments
+      {
+        CustomerId = customerId,
+        VendorId = venderId,
+        Comment = commentDto.Comment,
+        Rating = commentDto.Rating
+      };
+
       await _mongoDBService.CreateCommentAndRate(comment);
       return CreatedAtAction(nameof(Get), new { id = comment.Id }, comment);
     }
@@ -102,10 +138,16 @@ public class CommentController : ControllerBase
 
 
   [HttpPut("{id:length(24)}")]
-  public async Task<IActionResult> Update(string id, Comments updatedComment)
+  public async Task<IActionResult> Update(string id, CommentDto updatedCommentdto)
   {
     try
     {
+      var customerId = await AuthorizeCustomerAsync();
+      if (customerId == null)
+      {
+        return Unauthorized("Unauthorised User");
+      }
+
       var comment = await _mongoDBService.GetCommentAsync(id);
 
       if (comment is null)
@@ -113,7 +155,14 @@ public class CommentController : ControllerBase
         return NotFound(new { Message = "Comment not found" });
       }
 
-      updatedComment.Id = comment.Id;
+      var updatedComment = new Comments
+      {
+        Id = comment.Id,
+        CustomerId = customerId,
+        VendorId = comment.VendorId,
+        Comment = updatedCommentdto.Comment,
+        Rating = updatedCommentdto.Rating
+      };
 
       await _mongoDBService.UpdateCommentAsync(id, updatedComment);
 
@@ -135,6 +184,12 @@ public class CommentController : ControllerBase
   {
     try
     {
+      var customerId = await AuthorizeCustomerAsync();
+      if (customerId == null)
+      {
+        return Unauthorized("Unauthorised User");
+      }
+
       var isRemoved = await _mongoDBService.DeleteCommentAsync(id);
       if (!isRemoved)
       {
@@ -151,6 +206,111 @@ public class CommentController : ControllerBase
       return StatusCode(500, new { Message = "An unexpected error occurred", Error = error.Message });
     }
 
+  }
+
+  private async Task<string> AuthorizeVenderAsync()
+  {
+    // Validate token
+    var token = Request.Headers["Authorization"];
+    if (string.IsNullOrEmpty(token))
+    {
+      return null;
+    }
+
+    // Call JWT Service to validate the token
+    var tokenverify = JWTService.ValidateToken(token, _jwtSettings.SecurityKey);
+
+    // Check if the token is valid
+    if (tokenverify == null)
+    {
+      return null;
+    }
+
+    // Check if the user ID is in the token
+    var idClaim = tokenverify.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier);
+    if (idClaim == null)
+    {
+      return null;
+    }
+
+    // Check if the user is an vender
+    var VenderUser = await _mongoDBService.GetVendorByIdAsync(idClaim.Value);
+    if (VenderUser == null)
+    {
+      return null;
+    }
+
+    return idClaim.Value;
+  }
+
+  private async Task<string> AuthorizeCustomerAsync()
+  {
+    // Validate token
+    var token = Request.Headers["Authorization"];
+    if (string.IsNullOrEmpty(token))
+    {
+      return null;
+    }
+
+    // Call JWT Service to validate the token
+    var tokenverify = JWTService.ValidateToken(token, _jwtSettings.SecurityKey);
+
+    // Check if the token is valid
+    if (tokenverify == null)
+    {
+      return null;
+    }
+
+    // Check if the user ID is in the token
+    var idClaim = tokenverify.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier);
+    if (idClaim == null)
+    {
+      return null;
+    }
+
+    // Check if the user is an vender
+    var VenderUser = await _mongoDBService.GetCustomerByIdAsync(idClaim.Value);
+    if (VenderUser == null)
+    {
+      return null;
+    }
+
+    return idClaim.Value;
+  }
+
+  private async Task<string> AuthorizeAdminAsync()
+  {
+    // Validate token
+    var token = Request.Headers["Authorization"];
+    if (string.IsNullOrEmpty(token))
+    {
+      return null;
+    }
+
+    // Call JWT Service to validate the token
+    var tokenverify = JWTService.ValidateToken(token, _jwtSettings.SecurityKey);
+
+    // Check if the token is valid
+    if (tokenverify == null)
+    {
+      return null;
+    }
+
+    // Check if the user ID is in the token
+    var idClaim = tokenverify.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier);
+    if (idClaim == null)
+    {
+      return null;
+    }
+
+    // Check if the user is an vender
+    var VenderUser = await _mongoDBService.GetAdminByIdAsync(idClaim.Value);
+    if (VenderUser == null)
+    {
+      return null;
+    }
+
+    return idClaim.Value;
   }
 
 }
